+++ libs/base/src/code/util/mod.rs
@@ -12,6 +12,7 @@ mod _reexport; // reexport!, reexport_from!
 mod reexports; // re-exported macros from devela_base_macros
 
 mod cfg_if; // cfg_if!
+mod cfor; // cfor!
 mod deprecate; // deprecate_feature!
 mod include; // include_from!, mod_from!
 mod items; // items!, sf!
@@ -30,7 +31,7 @@ crate::items! { // structural access: _mods, _workspace_private, _all
 
     mod _mods {
         pub use super::{
-            cfg_if::*, deprecate::*, include::*, items::*, is::*,
+            cfg_if::*, cfor::*, deprecate::*, include::*, items::*, is::*,
             paste::*, r#const::*, reexports::*,
         };
     }