+++ libs/base/src/data/sort/definition.rs
@@ -1,10 +1,10 @@
-// devela::data::sort::definition
+// devela_base::data::sort::definition
 //
 //! Defines and documents [`Sort`].
 //
 
 #[doc = crate::TAG_NAMESPACE!()]
-/// Provides sorting methods for arrays and slices of `T`, some of them *const*.
+/// Provides sorting methods for arrays and slices of `T`.
 ///
 /// It implements the following methods for sorting exclusive slices:
 /// [`bubble`][Sort#bubble],
@@ -18,46 +18,42 @@
 /// [`quick_selection`][Sort#quick_selection],
 /// [`quick_shaker`][Sort#quick_shaker].
 ///
-/// # Features
-/// When the corresponding feature `_sort_f[32|64]` or `_sort_[iu][8|16|32|64|128]` is enabled,
-/// It implements the following *const* methods for sorting copied arrays of primitives:
-/// `bubble_array`, `insertion_array`, `selection_array`.
-/// In the case of floating-point primitives it uses total ordering.
-///
 /// # Examples
 /// Sort copied arrays of primitives:
 /// ```
-/// # use devela::Sort;
-/// # #[cfg(feature = "_sort_i32")]
-/// # {
+/// # use devela_base::Sort;
 /// let mut arr = [4i32, 7, -5, 1, -13, 0]; // signed primitives
 /// assert_eq![Sort(arr).bubble_array(),    [-13, -5, 0, 1, 4, 7]];
 /// assert_eq![Sort(arr).insertion_array(), [-13, -5, 0, 1, 4, 7]];
 /// assert_eq![Sort(arr).selection_array(), [-13, -5, 0, 1, 4, 7]];
-/// # }
 ///
-/// # #[cfg(feature = "_sort_u32")]
-/// # {
 /// let mut arr = [4u32, 7, 5, 1, 13, 0]; // unsigned primitives
 /// assert_eq![Sort(arr).bubble_array(),    [0, 1, 4, 5, 7, 13]];
 /// assert_eq![Sort(arr).insertion_array(), [0, 1, 4, 5, 7, 13]];
 /// assert_eq![Sort(arr).selection_array(), [0, 1, 4, 5, 7, 13]];
-/// # }
 ///
-/// # #[cfg(feature = "_sort_f32")]
-/// # {
 /// let mut arr = [4.01f32, 7.9, -5.4, 1.0, 0.0, -0.0]; // floating-point primitives
 /// assert_eq![Sort(arr).bubble_array(),    [-5.4, -0.0, 0.0, 1.0, 4.01, 7.9]];
 /// assert_eq![Sort(arr).insertion_array(), [-5.4, -0.0, 0.0, 1.0, 4.01, 7.9]];
 /// assert_eq![Sort(arr).selection_array(), [-5.4, -0.0, 0.0, 1.0, 4.01, 7.9]];
-/// # }
+///
+/// // compile-time friendly
+/// const SORTED: [f32; 6] = Sort([4.01f32, 7.9, -5.4, 1.0, 0.0, -0.0]).bubble_array();
+/// assert_eq![SORTED, [-5.4, -0.0, 0.0, 1.0, 4.01, 7.9]];
 /// ```
 ///
 /// # Performance
-/// The `_array` suffixed methods calls the [`cswap`] macro using the xor swap
+/// The `_array` suffixed methods calls the [`cswap!`] macro using the xor swap
 /// algorithm, except for the floting-point version which uses a temporary variable.
 ///
 /// [`cswap`]: crate::cswap
 #[repr(transparent)]
 #[derive(Debug)]
 pub struct Sort<T>(pub T);
+
+impl<T> Sort<T> {
+    ///
+    pub fn new(inner: T) -> Self {
+        Self(inner)
+    }
+}