+++ src/code/util/mod.rs
@@ -21,7 +21,6 @@ mod _use; // _use!
 mod asserts; // assertion macros
 mod cdbg; // cdbg!
 mod capture; // capture_[first|last|tail]!
-mod cfor; // cfor!
 mod ident; // ident_const_index!
 mod impl_trait; // impl_trait!
 mod maybe; // maybe!
@@ -41,7 +40,7 @@ reexports::items! { // structural access: _mods, _internals, _all, _always
 
     mod _mods {
         pub use super::{
-            asserts::_all::*, cdbg::*, capture::*, cfor::*, ident::*, impl_trait::*, maybe::*,
+            asserts::_all::*, cdbg::*, capture::*, ident::*, impl_trait::*, maybe::*,
             methods::*, reexports::*,
         };
         #[cfg(_bit··)]