+++ src/data/mod.rs
@@ -11,7 +11,6 @@
 
 mod absence; // NoData
 mod collection;
-mod sort;
 
 pub mod codec;
 pub mod iter;
@@ -32,7 +31,11 @@ crate::items! { // structural access: _mods, _pub_mods, _internals, _all, _alway
     pub use {_always::*, _pub_mods::*};
 
     mod _mods { #![allow(unused)]
-        pub use super::{absence::*, collection::*, sort::_all::*};
+        pub use super::{absence::*, collection::*};
+
+        pub use devela_base::Sort;
+        // #[cfg(feature = "alloc")]
+        // pub use devela_base_alloc::SortAlloc;
     }
     mod _pub_mods { #![allow(unused)]
         pub use super::{